@@ -3,64 +3,71 @@
   "version": "1.0.0",
   "description": "",
   "main": "metalsmith.js",
+  "browserslist": "> 0.25%, not dead",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "postinstall": "node tasks/postinstall.js",
     "lint": "eslint .",
     "start": "node tasks/dev.js",
-    "debug": "node --debug --inspect tasks/dev.js",
+    "dev": "node --inspect tasks/dev.js",
+    "debug": "node --inspect-brk tasks/dev.js",
     "build": "node metalsmith.js",
-    "postbuild": "./tasks/uglify build/js/common.js",
+    "postbuild": "uglifyjs --compress --mangle -o build/src/js/main.js -- build/src/js/main.js",
     "deploy": "node ./tasks/deploy.js",
     "deploy-locally": "node -r dotenv/config ./tasks/deploy.js"
   },
   "author": "Robert Preus-MacLaren",
-  "license": "GNUv3",
+  "license": "UNLICENSED",
   "dependencies": {
-    "autoprefixer": "8.6.2",
-    "babel-cli": "6.26.0",
-    "babel-preset-es2015": "6.24.1",
-    "babelify": "8.0.0",
-    "browserify": "16.2.2",
-    "cheerio": "0.22.0",
-    "circular-json": "0.5.4",
+    "@babel/core": "7.13.8",
+    "@babel/preset-env": "7.13.8",
+    "@babel/preset-es2015": "7.0.0-beta.53",
+    "autoprefixer": "10.2.4",
+    "babelify": "10.0.0",
+    "cheerio": "1.0.0-rc.5",
+    "circular-json": "0.5.9",
     "classlist-polyfill": "1.2.0",
-    "concurrently": "3.5.1",
-    "cssnano": "3.10.0",
-    "debounce": "1.1.0",
-    "dotenv": "6.0.0",
-    "eslint": "4.19.1",
+    "concurrently": "6.0.0",
+    "cssnano": "4.1.10",
+    "debounce": "1.2.0",
+    "dotenv": "8.2.0",
+    "eslint": "7.20.0",
     "factor-bundle": "2.5.0",
-    "foundation-sites": "6.4.3",
+    "foundation-sites": "6.3.1",
     "ftp": "0.3.10",
-    "ftp-deploy": "2.1.1",
-    "hamburgers": "0.9.3",
-    "handlebars": "4.0.11",
-    "headroom.js": "0.9.4",
-    "jquery": "3.3.1",
+    "ftp-deploy": "2.4.1",
+    "hamburgers": "1.1.3",
+    "handlebars": "4.7.7",
+    "headroom.js": "0.12.0",
+    "jquery": "3.5.1",
+    "jstransformer-handlebars": "1.1.0",
     "metalsmith": "2.3.0",
     "metalsmith-assets": "0.1.0",
-    "metalsmith-browserify": "1.0.0",
+    "metalsmith-browserify": "1.1.0",
     "metalsmith-collections": "0.9.0",
-    "metalsmith-discover-helpers": "0.1.0",
+    "metalsmith-discover-helpers": "0.1.1",
+    "metalsmith-discover-partials": "^0.1.2",
     "metalsmith-external-links": "0.1.2",
-    "metalsmith-layouts": "2.1.0",
-    "metalsmith-markdown": "0.2.2",
+    "metalsmith-layouts": "2.3.1",
+    "metalsmith-markdown": "1.3.0",
     "metalsmith-metadata-as-list": "0.1.0",
-    "metalsmith-permalinks": "0.5.0",
-    "metalsmith-postcss": "4.1.1",
-    "normalize.css": "8.0.0",
-    "postcss-import": "11.1.0",
+    "metalsmith-permalinks": "2.2.0",
+    "metalsmith-postcss": "4.2.0",
+    "normalize.css": "8.0.1",
+    "postcss": "8.2.6",
+    "postcss-import": "14.0.0",
     "progress-barjs": "2.2.1",
     "request-frame": "1.5.3",
-    "stat-mode": "0.2.2",
-    "uglify-js": "3.4.0"
+    "stat-mode": "1.0.0",
+    "uglify-js": "3.12.8"
   },
   "devDependencies": {
+    "babel-eslint": "^10.1.0",
     "blurmage": "1.0.2",
-    "body-parser": "1.18.3",
-    "decache": "4.4.0",
-    "express": "4.16.3",
+    "body-parser": "1.19.0",
+    "decache": "4.6.0",
+    "express": "4.17.1",
     "gaze": "1.1.3",
-    "reload": "2.2.2"
+    "reload": "3.1.1"
   }
 }
